WPT runner + diff output
